Add support for boolean values in MemoryStructureProxy (Issue #304)

+++ domino-jnx-api/src/main/java/com/hcl/domino/richtext/records/ViewmapRegionRecord.java
@@ -55,7 +55,7 @@ public interface ViewmapRegionRecord extends RichTextRecord<BSIG> {
   }
 
   @StructureGetter("LineColor")
-  short getLineColorRaw();
+  int getLineColorRaw();
   
   @StructureGetter("LineColor")
   Optional<StandardColors> getLineColor();
